call backend access before removing a file
this allows delete to focus on only removing existing files, with
the access call asserting file existence.

+++ src/ior.c
@@ -1272,13 +1272,15 @@ static void RemoveFile(char *testFileName, int filePerProc, IOR_param_t * test)
                         rankOffset = 0;
                         GetTestFileName(testFileName, test);
                 }
-                backend->delete(testFileName, test);
+                if (backend->access(testFileName, F_OK, test) == 0) {
+                        backend->delete(testFileName, test);
+                }
                 if (test->reorderTasksRandom == TRUE) {
                         rankOffset = tmpRankOffset;
                         GetTestFileName(testFileName, test);
                 }
         } else {
-                if (rank == 0) {
+                if ((rank == 0) && (backend->access(testFileName, F_OK, test) == 0)) {
                         backend->delete(testFileName, test);
                 }
         }
